+++ frontend/src/metabase/databases/components/DriverWarning/DriverWarning.tsx
